Move process filter to separated component
